Add PreserveMixin, and validate urls/emails.
Added PreserveMixin, because I like mixins. I mean I could have defined
a subclass of MappingSchema, but I think a mixin is more apropos.

Added validators for URL and emails. For some reason, colander.url
is a instance of the Regex type, and thus cannot have it's validation
message be changed, but colander.Email is a subclass of regex, which you
CAN change the validation message of. I'm not doing anything with the
validation messages, so it's not a big deal, just thought it was weird.
I mean if anything, I can just do `colander.Regex(colander.URL_REGEX, "my
message")`